requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod==1.5
python-dotenv
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp
opencv-python
ffmpeg-python
psutil